fixed split

+++ memmove.c
@@ -6,7 +6,7 @@
 /*   By: obult <obult@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/03 22:21:25 by obult         #+#    #+#                 */
-/*   Updated: 2020/11/04 14:10:25 by obult         ########   odam.nl         */
+/*   Updated: 2020/11/04 18:29:24 by obult         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int		i;
-	char	*csrc;
-	char	*cdest;
+	size_t		i;
+	char		*csrc;
+	char		*cdest;
 
 	i = 0;
 	csrc = (char *)src;
